develop: refactored test-suite
