+++ test/Control/Eff/Fresh/Test.hs
@@ -1,0 +1,31 @@
+{-# LANGUAGE FlexibleContexts, NoMonomorphismRestriction #-}
+{-# LANGUAGE TypeOperators, DataKinds #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TemplateHaskell #-}
+
+module Control.Eff.Fresh.Test (testGroups) where
+
+import Test.HUnit hiding (State)
+import Control.Eff
+import Control.Eff.Fresh
+import Control.Eff.Trace
+import Control.Monad (liftM2, msum, guard, mzero, mplus)
+import Data.OpenUnion
+import Utils
+
+import Test.Framework.TH
+import Test.Framework.Providers.HUnit
+
+testGroups = [ $(testGroupGenerator) ]
+
+case_Fresh_tfresh' :: Assertion
+case_Fresh_tfresh' = do
+  ((), actual) <- catchOutput tfresh'
+  assertEqual "Fresh: test"
+    (unlines ["Fresh 0", "Fresh 1"]) actual
+  where
+    tfresh' = runTrace $ flip runFresh' 0 $ do
+      n <- fresh
+      trace $ "Fresh " ++ show n
+      n <- fresh
+      trace $ "Fresh " ++ show n
