+++ test/Control/Eff/Writer/Strict/Test.hs
@@ -1,0 +1,24 @@
+{-# LANGUAGE FlexibleContexts, NoMonomorphismRestriction #-}
+{-# LANGUAGE TypeOperators, DataKinds #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TemplateHaskell #-}
+
+module Control.Eff.Writer.Strict.Test (testGroups) where
+
+import Test.HUnit hiding (State)
+import Control.Eff
+import Control.Eff.Writer.Strict
+import Control.Monad (liftM2, msum, guard, mzero, mplus)
+import Data.OpenUnion
+import Utils
+
+import Test.Framework.TH
+import Test.Framework.Providers.HUnit
+
+testGroups = [ $(testGroupGenerator) ]
+
+case_Strict1_Writer_runLastWriter :: Assertion
+case_Strict1_Writer_runLastWriter = let
+  ((), Just m) = run $ runLastWriter $ mapM_ tell [undefined, ()]
+  in
+   assertUndefined (m :: ())
